refactor: refactor constants and interface
